Create module for RL2
