+++ src/garage/tf/algos/rl2.py
@@ -1,9 +1,13 @@
-"""Wrapper class that is specifc for RL2."""
+"""Module for RL2.
+
+This module contains RL2Worker and the environment wrapper for RL2.
+"""
 import akro
 import gym
 import numpy as np
 
 from garage.envs.env_spec import EnvSpec
+from garage.sampler.worker import DefaultWorker
 
 
 class RL2Env(gym.Wrapper):
@@ -96,3 +100,60 @@ class RL2Env(gym.Wrapper):
 
         """
         return self._spec
+
+
+class RL2Worker(DefaultWorker):
+    """Initialize a worker for RL2.
+
+    In RL2, policy does not reset between trajectories in each meta batch.
+    Policy only resets once at the beginning of a trial/meta batch.
+
+    Args:
+        seed(int): The seed to use to intialize random number generators.
+        max_path_length(int or float): The maximum length paths which will
+            be sampled. Can be (floating point) infinity.
+        worker_number(int): The number of the worker where this update is
+            occurring. This argument is used to set a different seed for each
+            worker.
+        n_paths_per_trial (int): Number of trajectories sampled per trial/
+            meta batch. Policy resets in the beginning of a meta batch,
+            and obtain `n_paths_per_trial` trajectories in one meta batch.
+
+    Attributes:
+        agent(Policy or None): The worker's agent.
+        env(gym.Env or None): The worker's environment.
+
+    """
+
+    def __init__(
+            self,
+            *,  # Require passing by keyword, since everything's an int.
+            seed,
+            max_path_length,
+            worker_number,
+            n_paths_per_trial=2):
+        self._n_paths_per_trial = n_paths_per_trial
+        super().__init__(seed=seed,
+                         max_path_length=max_path_length,
+                         worker_number=worker_number)
+
+    def start_rollout(self):
+        """Begin a new rollout."""
+        self._path_length = 0
+        self._prev_obs = self.env.reset()
+
+    def rollout(self):
+        """Sample a single rollout of the agent in the environment.
+
+        Returns:
+            garage.TrajectoryBatch: The collected trajectory.
+
+        """
+        self.agent.reset()
+        for _ in range(self._n_paths_per_trial):
+            self.start_rollout()
+            while not self.step_rollout():
+                pass
+        self._agent_infos['batch_idx'] = np.full(len(self._rewards),
+                                                 self._worker_number)
+        return self.collect_rollout()
