--- src/garage/tf/algos/rl2/__init__.py
@@ -1,5 +1,0 @@
-"""Module for RL2."""
-from garage.tf.algos.rl2.rl2_worker import RL2Worker
-from garage.tf.algos.rl2.rl2env import RL2Env
-
-__all__ = ['RL2Env', 'RL2Worker']
